+++ butterknife-compiler/src/main/java/butterknife/internal/FieldBitmapBinding.java
@@ -1,0 +1,19 @@
+package butterknife.internal;
+
+final class FieldBitmapBinding {
+  private final int id;
+  private final String name;
+
+  FieldBitmapBinding(int id, String name) {
+    this.id = id;
+    this.name = name;
+  }
+
+  public int getId() {
+    return id;
+  }
+
+  public String getName() {
+    return name;
+  }
+}