+++ butterknife-sample/src/main/java/com/example/butterknife/SimpleAdapter.java
@@ -7,7 +7,7 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
 import butterknife.ButterKnife;
-import butterknife.FindView;
+import butterknife.Bind;
 
 public class SimpleAdapter extends BaseAdapter {
   private static final String[] CONTENTS =
@@ -51,9 +51,9 @@ public class SimpleAdapter extends BaseAdapter {
   }
 
   static class ViewHolder {
-    @FindView(R.id.word) TextView word;
-    @FindView(R.id.length) TextView length;
-    @FindView(R.id.position) TextView position;
+    @Bind(R.id.word) TextView word;
+    @Bind(R.id.length) TextView length;
+    @Bind(R.id.position) TextView position;
 
     ViewHolder(View view) {
       ButterKnife.bind(this, view);